Various renaming for consistency
