+++ docs/api/manson_power_supply/api.md
@@ -2,5 +2,5 @@
 See the [how-to guide](../../devices/misc/manson.md) to configure a MansonPowerSupply object.
 
 ```{eval-rst}
-.. openapi:: ./manson.yml
+.. openapi:: ./manson_power_supply.yml
 ```
