+++ docs/api/manson_power_supply/manson_power_supply.yml
