+++ docs/api/phidget_p_sensor/api.md
@@ -1,0 +1,7 @@
+# Phidget Pressure Sensor
+
+See the [how-to guide](../../devices/sensors/phidget_pressure_sensor.md) to configure a PhidgetPressureSensor object.
+
+```{eval-rst}
+.. openapi:: ./phidget_pressure_sensor.yml
+```
