+++ docs/api/phidget_p_sensor/phidget_pressure_sensor.yml
@@ -1,0 +1,142 @@
+openapi: 3.0.2
+info:
+  title: Flowchem - devices
+  description: Flowchem is a python library to control a variety of instruments commonly
+    found in chemistry labs.
+  license:
+    name: MIT License
+    url: https://opensource.org/licenses/MIT
+  version: 0.1.0a3
+paths:
+  /hplc/run:
+    put:
+      tags:
+      - hplc
+      - hplc
+      summary: Run
+      description: 'Run one analysis on the instrument. The sample name has to be
+        set in advance via sample-name.
+
+
+        Note that it takes at least 2 sec until the run actually starts (depending
+        on instrument configuration).
+
+        While the export of the chromatogram in e.g. ASCII format can be achieved
+        programmatically via the CLI, the best
+
+        solution is to enable automatic data export for all runs of the HPLC as the
+        chromatogram will be automatically
+
+        exported as soon as the run is finished.'
+      operationId: run_hplc_run_put
+      responses:
+        '200':
+          description: Successful Response
+          content:
+            application/json:
+              schema: {}
+  /hplc/method:
+    put:
+      tags:
+      - hplc
+      - hplc
+      summary: Set Method
+      description: 'Sets the HPLC method (i.e. a file with .MET extension) to the
+        instrument.
+
+
+        Make sure to select ''Send Method to Instrument'' option in Method Sending
+        Options dialog in System Configuration.'
+      operationId: set_method_hplc_method_put
+      parameters:
+      - required: true
+        schema:
+          title: Method Name
+          type: string
+        name: method_name
+        in: query
+      responses:
+        '200':
+          description: Successful Response
+          content:
+            application/json:
+              schema: {}
+        '422':
+          description: Validation Error
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/HTTPValidationError'
+  /hplc/sample-name:
+    put:
+      tags:
+      - hplc
+      - hplc
+      summary: Set Sample Name
+      description: Sets the name of the sample for the next run.
+      operationId: set_sample_name_hplc_sample_name_put
+      parameters:
+      - required: true
+        schema:
+          title: Sample Name
+          type: string
+        name: sample_name
+        in: query
+      responses:
+        '200':
+          description: Successful Response
+          content:
+            application/json:
+              schema: {}
+        '422':
+          description: Validation Error
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/HTTPValidationError'
+  /hplc/exit:
+    put:
+      tags:
+      - hplc
+      - hplc
+      summary: Exit
+      description: Exit Clarity Chrom.
+      operationId: exit_hplc_exit_put
+      responses:
+        '200':
+          description: Successful Response
+          content:
+            application/json:
+              schema: {}
+components:
+  schemas:
+    HTTPValidationError:
+      title: HTTPValidationError
+      type: object
+      properties:
+        detail:
+          title: Detail
+          type: array
+          items:
+            $ref: '#/components/schemas/ValidationError'
+    ValidationError:
+      title: ValidationError
+      required:
+      - loc
+      - msg
+      - type
+      type: object
+      properties:
+        loc:
+          title: Location
+          type: array
+          items:
+            anyOf:
+            - type: string
+            - type: integer
+        msg:
+          title: Message
+          type: string
+        type:
+          title: Error Type
+          type: string
