+++ docs/devices/sensors/phidget_pressure_sensor.md
