+++ docs/devices/supported_devices.md
@@ -10,6 +10,7 @@ pumps/index.md
 valves/index.md
 temperature/index.md
 pressure/index.md
+sensors/index.md
 misc/index.md
 
 ```
